+++ src/TypedTables.jl
@@ -4,14 +4,17 @@ using NullableArrays
 
 export Field, DefaultKey, FieldIndex, Cell, Column, Row, Table, DenseTable, KeyTable, DenseKeyTable
 
-export rename, name, eltypes, field, index, key, keyname, ncol, nrow, head, tail, samefield, samefields
+export rename, name, eltypes, field, index, key, keyname, ncol, nrow, head, tail, samefield, samefields, storagetype, storagetypes
 
 export join, unique!
+
+export readtable, readtable!, writetable
 
 export  @field, @index, @cell, @row, @column, @table
 
 import Base.(==)
 
+# TODO union!, intersect! and setdiff!
 # TODO dplyr funcitons (should we use same names or julia interfaces or both??)
 #      - filter (slice is indexing)
 #      - arrange (sorting)
@@ -22,14 +25,11 @@ import Base.(==)
 #      - sample_n and sample_frac (could use rand() interface I think)
 #      - left_join, right_join, full_join (aka outer_join), semi_join, anti_join
 # TODO Load from an object with a dataframe interface
-# TODO Input and output to CSV, DLM
-# TODO similarly to ==, push!, etc think about vcat, too, for matching orderings? At the moment there is a difference between how vcat and a potential `append()` function would behave...
 # TODO Figure out namespace issues for head and tail (maybe both are in Base for Julia 0.5?)
-# TODO sub for generating a sub-table (i.e. simply a table with different StorageTypes)
-# TODO possibly implement subtable type (with no ability to push! or change rows, though can setindex!) - is this only necessary for Row?
+# TODO sub/slice for generating a sub-table (i.e. simply a table with different StorageTypes)
+# TODO possibly implement subtable type (with no ability to insert/delete rows, though can setindex!) - is this only necessary for Row?
 # TODO fix TableKey so it always references the length of the current table, not its parent
 # TODO Somehow make sense of the key mess for joins
-# TODO Set opertations like unique/unique! and union, intersect, setdiff
 # TODO Unary relational operations like select, project, etc (see dplyr for inspiration)
 # TODO Binary relational operations like left_join, etc.
 # TODO other types of computed joins, map, do, etc