+++ test/runtests.jl
@@ -16,8 +16,9 @@ include("test_column.jl")
 include("test_fieldindex.jl")
 include("test_row.jl")
 include("test_table.jl")
+
 include("test_join.jl")
 include("test_setalgorithms.jl")
-
+include("test_IO.jl")
 
 end