+++ NEWS.md
@@ -5,7 +5,9 @@ Note: minor changes and bug fixes may be ommited from this document.
 ### 0.0.3
 
 - Set operations: `unique()`, `unique!()`, `uniqueind()`, `groupinds()`, `union()`, `intersect()` and `setdiff()`.
-- Ability to import a table from a `DataFrame` or file `IOStream` using `readtable`
+- Ability to import a table from a `DataFrame` or file `IOStream` using `readtable()`
+- Ability to write a table to CSV/DLM format using `writetable()`
+- `vcat` for reordered indices, like already possible for `append!`
 
 ### 0.0.2